backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Supabase
supabase>=2.0.0
postgrest>=0.13.0

# ML & Embeddings
sentence-transformers>=2.3.0
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0

# Google Gemini
google-generativeai>=0.3.0

# Groq AI
groq>=0.4.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0

# Rate Limiting
slowapi>=0.1.9


